Rails MarkIt 7: User authentication

+++ app/controllers/bookmarks_controller.rb
@@ -1,5 +1,9 @@
 class BookmarksController < ApplicationController
+  before_action :authenticate_user!
+  rescue_from Pundit::NotAuthorizedError, with: :user_not_authorized
+
   def show
+    @topic = Topic.find(params[:topic_id])
   end
 
   def new
@@ -9,7 +13,8 @@ class BookmarksController < ApplicationController
 
   def create
     @topic = Topic.find(params[:topic_id])
-    @bookmark = @topic.bookmarks.build(bookmark_params)
+    @bookmark = @topic.bookmarks.create(bookmark_params)
+    @bookmark.user = current_user
 
     if @bookmark.save
       flash[:notice] = "Bookmark was saved successfully."
@@ -25,12 +30,14 @@ class BookmarksController < ApplicationController
   end
 
   def update
+    @topic = Topic.find(params[:topic_id])
     @bookmark = Bookmark.find(params[:id])
     @bookmark.assign_attributes(bookmark_params)
+    authorize @bookmark
 
     if @bookmark.save
       flash[:notice] = "Bookmark was updated successfully."
-      redirect_to [@bookmark.topic]
+      redirect_to topic_path(@topic)
     else
       flash.now[:alert] = "There was an error saving the post. Please try again."
       render :edit
@@ -38,14 +45,16 @@ class BookmarksController < ApplicationController
   end
 
   def destroy
+    @topic = Topic.find(params[:topic_id])
     @bookmark = Bookmark.find(params[:id])
+    authorize @bookmark
 
     if @bookmark.destroy
       flash[:notice] = "\"#{@bookmark.url}\" was deleted successfully."
-      redirect_to [@bookmark.topic]
+      redirect_to topic_path(@topic)
     else
       flash.now[:alert] = "There was an error deleting the bookmark."
-      render :show
+      redirect_to topic_path(@topic)
     end
   end
 
@@ -53,4 +62,10 @@ class BookmarksController < ApplicationController
   def bookmark_params
     params.require(:bookmark).permit(:url)
   end
+
+  def user_not_authorized
+    flash[:alert] = "You are not authorized to perform this action."
+    redirect_to(request.referrer || root_path)
+  end
+
 end
